Rolling back utility imports, fixing up env.evaluate
Importing via the utility file was causing env.evaluate to fail with
non-DistModel models.  I’m rolling it back to work on other stuff.

On the bright side, I think I’ve figured out that predict() can be used
with all models just by including BOTH an x and a newdata argument; all
of the predict functions we’re using take a …, which means they’ll just
ignore the one they don’t use.  Seems to be working for the moment,
anyway.