Comments for ENMeval
No actual changes to functionality, but I commented some code in the
glm function that partitions using two of the methods in ENMeval (block
and checkerboard 1).  This is just sort of a mental bookmark since it’s
the end of the week and I am going home to take a nap.