Added check.env
I've just added a new function that checks env layers to make sure they're the same extent, and then sets layers so that an NA in a cell in any layer is propagated to the others.  This is intended to be run before any of the background generating functions, to make sure that you're not drawing background from cells that have NA for one or more variables.