Brought combine.species up to date for terra
We were sort of trying to have our cake and eat it too - accepting both rasters and terra SpatRasters, and that was screwing a lot of things up.  Now I'm just going to require the new data types and set combined types to NA with a warning if we don't get them.