Updated bias layer sampling
Terra doesn't behave very usefully when sampling with replacement when the number of points is greater than the number of cells OR when sampling with replacement using weights from a SpatRaster.  We need to do both so we had hand-rolled a solution.  This fix wasn't really ideal, though, so I'm hoping this fixed version will work better.